backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
langchain>=0.1.16
langchain-community>=0.0.34
langchain-core>=0.1.45
langchain-google-genai>=1.0.3
langchain-text-splitters>=0.0.1
faiss-cpu>=1.8.0
pdfplumber>=0.11.0
google-generativeai>=0.5.2
pydantic>=2.7.0
langchain-groq